Add CI pipeline to update archive
This commit adds a pipeline (Jenkinsfile) and some auxilliary scripts
(under `tools/ci/`) to build the `btfhub` tool, generate BTFs and push
to the `btfhub-archive` repository.

This pipeline will be scheduled daily, running between 0:00-6:59, as
determined by the CI system.

Due to repository restrictions, the `master` branch cannot be directly
updated by this pipeline; the `next` branch will be updated and a PR
will be automatically created.