Revert changes of credentials to authenticate with Gitee
In PR !4, the credentials used for `git push` and sending requests to
Gitee APIs have been switched to @I-am-a-robot, but this requires to
grant @I-am-a-robot committer priviledges, which has not been approved
yet [1].

In this commit, the user used for authentication with Gitee is switched
back to @YangHanlin to avoid errors in CI builds. This commit can be
reverted once [1] is merged.

[1] https://gitee.com/openeuler/community/pulls/4942